+++ sdrs/run.sh
@@ -1,15 +1,22 @@
 #!/bin/bash
 
+if [[ -z "${DOCKER_USER}" ]]; then
+    echo "Please set DOCKER_USER environment variable!"
+    exit -1
+fi
+
+mkdir -p /tmp/dockershared
+
 docker rm sdrs
 
 docker run \
     --name sdrs \
     -it \
-    --user $USER \
+    --user $DOCKER_USER \
     --privileged \
     -v /dev/bus/usb:/dev/bus/usb \
     -e DISPLAY=$DISPLAY \
     -v /tmp/.X11-unix:/tmp/.X11-unix \
-    --add-host=host.docker.internal:host-gateway \
+    -v /tmp/dockershared:/home/$DOCKER_USER/dockershared \
     sdrs \
     bash
